Fixes from PR review
